pdf-service/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# File handling
aiofiles==23.2.1
python-multipart==0.0.6

# PDF processing
PyMuPDF<1.21.0
pdf2image==1.16.3
Pillow==10.1.0

# OCR and image processing
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python<=4.6.0.66
numpy>=1.21.0,<1.25.0

# Image processing dependencies (poppler-utils is installed as system dependency in Dockerfile)

# Async processing (optional)
celery==5.3.4
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0 